worker/requirements.txt
playwright==1.40.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.6.4
cryptography==41.0.7
pydantic==2.5.0
tenacity==8.2.3
